Lots of optimizations and sanity-checks

+++ deepdream/deepdream.py
@@ -1,6 +1,5 @@
 import caffe
 from google.protobuf import text_format
-from IPython.display import clear_output
 import numpy as np
 import os
 import PIL.Image
@@ -15,36 +14,40 @@ def deprocess(net, img):
     return np.dstack((img + net.transformer.mean["data"])[::-1])
 
 
-def make_step(net, step_size=1.5, end="inception_4c/output", jitter=32, clip=True):
+def objective_L2(dst):
+    dst.diff[:] = dst.data
+
+
+def make_step(net, step_size=1.5, end="inception_4c/output", jitter=32, clip=True, objective=objective_L2):
     """Basic gradient ascent step."""
 
-    src = net.blobs["data"]  # input image is storred in Net's 'data' blob
+    src = net.blobs['data']  # input image is stored in Net's 'data' blob
     dst = net.blobs[end]
 
     ox, oy = np.random.randint(-jitter, jitter + 1, 2)
     src.data[0] = np.roll(np.roll(src.data[0], ox, -1), oy, -2)  # apply jitter shift
 
     net.forward(end=end)
-    dst.diff[:] = dst.data  # specify the optimiation objective
+    objective(dst)  # specify the optimization objective
     net.backward(start=end)
     g = src.diff[0]
-    # apply normaized ascent step to the input image
+    # apply normalized ascent step to the input image
     src.data[:] += step_size / np.abs(g).mean() * g
 
-    src.data[0] = np.roll(np.roll(src.data[0], -ox, -1), -oy, -2)  # unshift image
+    src.data[0] = np.roll(np.roll(src.data[0], -ox, -1), -oy, -2)  # un-shift image
 
     if clip:
-        bias = net.transformer.mean["data"]
+        bias = net.transformer.mean['data']
         src.data[:] = np.clip(src.data, -bias, 255 - bias)
 
 
-def deepdream(net, base_img, iter_n=10, octave_n=4, octave_scale=1.4, end="inception_4c/output", clip=True, **step_params):
+def deepdream(net, base_img, iter_n=10, octave_n=4, octave_scale=1.4, end='inception_4c/output', clip=True, **step_params):
     # prepare base images for all octaves
     octaves = [preprocess(net, base_img)]
     for i in xrange(octave_n - 1):
         octaves.append(nd.zoom(octaves[-1], (1, 1.0 / octave_scale, 1.0 / octave_scale), order=1))
 
-    src = net.blobs["data"]
+    src = net.blobs['data']
     detail = np.zeros_like(octaves[-1])  # allocate image for network-produced details
     for octave, octave_base in enumerate(octaves[::-1]):
         h, w = octave_base.shape[-2:]
@@ -57,14 +60,6 @@ def deepdream(net, base_img, iter_n=10, octave_n=4, octave_scale=1.4, end="incep
         src.data[0] = octave_base + detail
         for i in xrange(iter_n):
             make_step(net, end=end, clip=clip, **step_params)
-
-            # visualization
-            vis = deprocess(net, src.data[0])
-            if not clip:  # adjust image contrast if clipping is disabled
-                vis = vis * (255.0 / np.percentile(vis, 99.98))
-            # showarray(vis)
-            print octave, i, end, vis.shape
-            clear_output(wait=True)
 
         # extract details produced on the current octave
         detail = src.data[0] - octave_base
@@ -79,12 +74,13 @@ def dream_about(input_image_path, end_layer, width, output_folder):
 
     # Patching model to be able to compute gradients.
     # Note that you can also manually add "force_backward: true" line to "deploy.prototxt".
-    model = caffe.io.caffe_pb2.NetParameter()
-    with open(net_fn) as fp:
-        text_format.Merge(fp.read(), model)
-    model.force_backward = True
-    with open("tmp.prototxt", "w") as fp:
-        fp.write(str(model))
+    if not os.path.exists("tmp.prototxt"):
+        model = caffe.io.caffe_pb2.NetParameter()
+        with open(net_fn) as fp:
+            text_format.Merge(fp.read(), model)
+        model.force_backward = True
+        with open("tmp.prototxt", "w") as fp:
+            fp.write(str(model))
 
     net = caffe.Classifier("tmp.prototxt", param_fn,
                            mean=np.float32([104.0, 116.0, 122.0]),  # ImageNet mean, training set dependent
@@ -92,22 +88,18 @@ def dream_about(input_image_path, end_layer, width, output_folder):
 
     # a couple of utility functions for converting to and from Caffe's input image layout
 
-    maxwidth = width
     img = PIL.Image.open(input_image_path)
-    width = img.size[0]
 
-    if width > maxwidth:
-        wpercent = (maxwidth / float(img.size[0]))
+    # if image exceeds maxwidth, scale using PIL.Image.ANTIALIAS
+    if img.size[0] > width:
+        wpercent = (width / float(img.size[0]))
         hsize = int((float(img.size[1]) * float(wpercent)))
-        img = img.resize((maxwidth, hsize), PIL.Image.ANTIALIAS)
+        img = img.resize((width, hsize), PIL.Image.ANTIALIAS)
 
+    # convert to numpy float32 and call actual deepdream function
     img = np.float32(img)
-
-    frame = img
-
-    frame = deepdream(net, frame, end=end_layer)
-
-    output_path = os.path.join(output_folder, os.path.basename(input_image_path))
+    frame = deepdream(net, img, end=end_layer)
 
     # save image to output path
+    output_path = os.path.join(output_folder, os.path.basename(input_image_path))
     PIL.Image.fromarray(np.uint8(frame)).save(output_path)
